Added Critical Force. Changed weight reading to float. Changed button enabling.
Added critical force pages.
Fixed export.
Changed TimestampedWeight to hold a float value.
DataCollector now returns a float instead of 100x scale reading as that's what it reports. For example if the scale reads 27.64 it'll emit 2764 as the reading.
Changes button enabling/disabling when in a live session. Start will now enable stop. Stop will disable start and enable save/export.

+++ app/src/main/java/com/example/dyna/TimestampedWeight.java
@@ -7,13 +7,13 @@ public class TimestampedWeight implements Serializable {
         return timestamp;
     }
 
-    public int getWeight() {
+    public float getWeight() {
         return weight;
     }
 
     private final long timestamp;
-    private final int weight;
-    public TimestampedWeight(int w){
+    private final float weight;
+    public TimestampedWeight(float w){
         this.timestamp = System.currentTimeMillis();
         this.weight = w;
     }
